more edit to sintegration for scene2d
now Z index is float, made a class for custom Actor Zcomparator

+++ src/main/java/org/gattolfo/engen/scene2d/ZActor.java
@@ -3,13 +3,13 @@ package org.gattolfo.engen.scene2d;
 import com.badlogic.gdx.scenes.scene2d.Actor;
 
 public class ZActor extends Actor implements Ztable{
-    private int z = 0;
+    private float z = 0;
 
-    public int getZ() {
+    public float getZ() {
         return z;
     }
 
-    public void setZ(int z) {
+    public void setZ(float z) {
         this.z = z;
     }
 }
