+++ src/main/java/org/gattolfo/engen/scene2d/Ztable.java
@@ -1,6 +1,6 @@
 package org.gattolfo.engen.scene2d;
 
 public interface Ztable {
-    public int getZ();
-    public void setZ(int z);
+    public float getZ();
+    public void setZ(float z);
 }
